% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interactions.R
\name{get_interactions}
\alias{get_interactions}
\title{Query interactions between drugs, targets, and diseases from Open Targets API}
\usage{
get_interactions(query_id, id_type)
}
\arguments{
\item{query_id}{A string representing the unique identifier for the query. This could be
a ChEMBL ID (for drugs), Ensembl ID (for targets), or EFO ID (for diseases).}

\item{id_type}{A string representing the type of the identifier. Must be one of "drug",
"target", or "disease".}
}
\value{
A data frame containing interaction details between drugs, targets, and diseases, including:
\describe{
\item{approvedName}{The approved name of the drug or target.}
\item{approvedSymbol}{The approved symbol for the target.}
\item{targetId}{The target identifier.}
\item{diseaseId}{The disease identifier.}
\item{drugId}{The drug identifier.}
}
If the query fails, the function returns a list with an error message.
}
\description{
This function queries the Open Targets Platform to retrieve interactions between drugs,
targets, and diseases based on the provided identifier and type (drug, target, or disease).
The function sends a GraphQL query to the Open Targets API and returns relevant interaction data.
}
\details{
The function sends a POST request to the Open Targets API with the appropriate GraphQL query
based on the provided \code{id_type}. The function handles the query results and returns relevant
data about the interactions between drugs, targets, and diseases.

The function supports a maximum of 10,000 results due to API limitations. If an invalid
\code{id_type} is provided, the function will return an error message.
}
\examples{
\dontrun{
# Query interactions for a drug with ChEMBL ID
result <- get_interactions("CHEMBL25", "drug")

# Query interactions for a target with Ensembl ID
result <- get_interactions("ENSG00000169083", "target")

# Query interactions for a disease with EFO ID
result <- get_interactions("EFO_0000222", "disease")
}

}
